Fix approximate count double test
Use DoubleType instead of BigintType

+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestApproximateCountDoubleAggregation.java
@@ -20,7 +20,7 @@ import com.google.common.base.Predicates;
 import java.util.List;
 
 import static com.facebook.presto.operator.aggregation.ApproximateCountColumnAggregations.DOUBLE_APPROXIMATE_COUNT_AGGREGATION;
-import static com.facebook.presto.spi.type.BigintType.BIGINT;
+import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 
 public class TestApproximateCountDoubleAggregation
         extends AbstractTestApproximateAggregationFunction
@@ -28,7 +28,7 @@ public class TestApproximateCountDoubleAggregation
     @Override
     protected Type getType()
     {
-        return BIGINT;
+        return DOUBLE;
     }
 
     @Override
